feat: remove dir if setup error.
